Add files via upload
Poster Issued by the "Central Historical Commission at the Central Committee of Liberated Jews in the American Zone of Occupation in Germany" – Stefania Scagliola has uploaded this image, even though we have no copyright, because the page on Ranke is from an auction site that may disappear. In this case we might be able to use this image, as it is hidden in the interactive animation.